Fix Various TransportReassembly Compile Warnings

+++ dds/DCPS/transport/framework/ReceivedDataSample.cpp
@@ -158,8 +158,8 @@ ReceivedDataSample::get_fragment_range(FragmentNumber start_frag, FragmentNumber
   result.header_ = header_;
 
   const size_t fsize = static_cast<size_t>(fragment_size_);
-  const size_t start_offset = start_frag * fsize;
-  const size_t end_offset = end_frag == INVALID_FRAGMENT ? std::numeric_limits<size_t>::max() : (end_frag + 1) * fsize - 1;
+  const size_t start_offset = static_cast<size_t>(start_frag) * fsize;
+  const size_t end_offset = end_frag == INVALID_FRAGMENT ? std::numeric_limits<size_t>::max() : static_cast<size_t>(end_frag + 1) * fsize - 1;
 
   size_t current_offset = 0;
